refactor: mvc의 각 기능을 폴더로 구분
